Extract app route definition to its own file
